// src/domain/enums/PowerUpType.ts
export enum PowerUpType {
  ArmorRegeneration = 'Armor Regeneration',
  BattleSuit = 'Battle Suit',
  Doubler = 'Doubler',
  Flight = 'Flight',
  Guard = 'Guard',
  Haste = 'Haste',
  Invisibility = 'Invisibility',
  Invulnerability = 'Invulnerability',
  OtherPowerup = 'Other powerup',
  QuadDamage = 'Quad Damage',
  Regeneration = 'Regeneration'
}