// src/domain/enums/WeaponType.ts
export enum WeaponType {
  Bfg = 'BFG',
  ChainGun = 'Chain Gun',
  Gauntlet = 'Gauntlet',
  GrenadeLauncher = 'Grenade Launcher',
  HeavyMachineGun = 'Heavy Machine Gun',
  LightningGun = 'Lightning Gun',
  MachineGun = 'Machine Gun',
  Nailgun = 'Nailgun',
  OtherWeapon = 'Other weapon',
  PlasmaGun = 'Plasma Gun',
  ProximityLauncher = 'Proximity Launcher',
  Railgun = 'Railgun',
  RocketLauncher = 'Rocket Launcher',
  Shotgun = 'Shotgun'
}